Rename src/fml.js -> src/client/forgeHandshake.js
